fix: add OnFallback type to GetRenderHandlerOptions
